devtools → remotes, syntax highlighting
Depending how fresh the user's library is, there may be quite a lot to install before the user can even call `devtools::install_github`. It would be less burdensome for the user to install the dependency-less {remotes} just to install this package from GitHub. Also, {remotes} takes care of installing dependencies in DESCRIPTION, so the call can be simplified.

Also I've specified which language each code block is in (where possible) to use GitHub's automatic syntax highlighting for R and Stan.